GO-2372 Update sys rels on ReindexMarketplace
